+ saturation levels. Big step. removed the affectMaterials structures in the SaturationEffect structures. it was unnecessary.
There are most likely some bugs with this (I am writing this at 2am in the morning)
but the system seems to be up and working!
you now require a certain level of saturation to achieve your affectMat.

also, now there are no AffectMats in the satEffect structure. you simply have to describe what material triggers your affectMaterial in the element of the affectMaterial structure called "satNeeded". it is all there in the comments.