fixed a crash bug, make the clear feature (c-key) clear all the cell data (not just the current screen cells), and made random randomize all cell data.
I added a few ideas to the features-ideas.txt doc.
The crash bug was that when you zoomed out, the current view of the cell data would exceed the bounds of the array.
basically, you would try to look at a point outside of the valid data and the program would shut down to avoid reading/writing data that it didn't have permissions to read/write. Basically a SEGFAULT error.