Started work on gravity. simple straight-down gravity is working fine. However, slope gravity still needs to be re-worked.
Materials are being lost in gravity calculations and I need to find out why. materials are being lost in both gravity < 0 and gravity > 0 cases for the gravity evaluation code of the cell_engine.
I'm considering reworking the code. The slope gravity was only ever half-done anyway. Half-baked, as it were.
I only allowed materials to fall down slopes through the air. They need to be able to flow into other materials as well (fluids and solids flow into lighter fluids). Also, I have some ideas for making a more fair gravity evaluation function. But they are too complex to write in the margin of this commit.